+++ src/components/index.ts
@@ -3,7 +3,7 @@ import Button from "@components/commons/Button/Button";
 import Header from "@components/layouts/Header/Header";
 import TextField from "@components/commons/TextField/TextField";
 import PotCard from "./cards/PotCard/PotCard";
-import CategoryButton from "./commons/Badge/CategoryButton/CategoryButton";
+import CategoryButton from "./commons/Button/CategoryButton/CategoryButton";
 import TaskCard from "./cards/TaskCard/TaskCard";
 import MemberGroup from "./commons/Badge/MemberGroup/MemberGroup";
 import StateBadge from "./commons/Badge/StateBadge/StateBadge";
@@ -11,11 +11,11 @@ import DdayBadge from "./commons/Badge/DdayBadge/DdayBadge";
 import Badge from "./commons/Badge/Badge";
 import SearchInput from "./commons/SearchInput/SearchInput";
 import CheckBox from "./commons/Button/CheckBox/CheckBox";
-import PotButton from "./commons/PotButton/PotButton";
+import PotButton from "./commons/Button/PotButton/PotButton";
 import ExplainModal from "./commons/Modal/ExplainModal/ExplainModal";
 import ArrowButton from "./commons/Button/ArrowButton/ArrowButton";
 import Dropdown from "./commons/Dropdown/Dropdown";
-import FloatingButton from "./commons/FloatingButton/FloatingButton";
+import FloatingButton from "./commons/Button/FloatingButton/FloatingButton";
 import Modal from "./commons/Modal/Modal";
 import MyFeedDropdown from "./commons/Dropdown/MyFeedDropdown/MyFeedDropdown";
 import PotInformation from "./commons/PotInformation/PotInformation";
@@ -24,7 +24,7 @@ import FinishedPotCard from "./cards/FinishedPotCard/FinishedPotCard";
 import InformationPopper from "./commons/InformationPopper/InformationPopper";
 import OnGoingPotCard from "./cards/OnGoingPotCard/OnGoingPotCard";
 import PartRecruitment from "./commons/PartRecruitment/PartRecruitment";
-import PostForm from "./commons/PostFoam/PostFoam";
+import PostForm from "./commons/PostForm/PostForm";
 
 export {
   Modal,
